moved mremap mapper to a separate file
